Enum jb_lib::tool::kind::Kind

pub enum Kind {

    IntelliJIdeaUltimate,
    IntelliJIdeaCommunity,
    PyCharmProfessional,
    PyCharmCommunity,
    PhpStorm,
    GoLand,
    Rider,
    CLion,
    ClionNova,
    RustRover,
    WebStorm,
    RubyMine,
    DataGrip,
    DataSpell,
    Fleet,
    Aqua,
    Writerside,
    DotMemory,
    DotTrace,
    MPS,
    Space,
    Gateway,

}

Tool kind

This enum does not contain any information. It only represents which kind of tool it is.

Variants

IntelliJIdeaUltimate

IntelliJIdeaCommunity

PyCharmProfessional

PyCharmCommunity

PhpStorm

GoLand

Rider

CLion

ClionNova

RustRover

WebStorm

RubyMine

DataGrip

DataSpell

Fleet

Aqua

Writerside

DotMemory

DotTrace

MPS

Space

Gateway

Implementations

impl Kind

pub fn list() -> &'static [Self]

Returns a list of all tool kinds.

This is used for display purposes and to loop over all tool kinds. The list is static and does not require any allocations.

pub fn as_code(&self) -> &str

Returns the code for this tool kind.

This is used to fetch the releases for this tool from the JetBrains website.

pub fn as_str(&self) -> &'static str

Returns the string representation of this tool kind.

This is used to determine the directory name for the tool as well as the arguments for the CLI.

pub fn pretty(&self) -> &'static str

Returns the pretty name for this tool kind.

This is used for display purposes.

pub fn src_name(&self) -> &str

Returns the source name for this tool kind.

This is used to determine the binary name for the tool, as some tools may have the same binary name while being different tools. (e.g. idea for IntelliJ IDEA Ultimate and IntelliJ IDEA Community)

pub fn default_type(&self) -> Type

Returns the default release type for this tool kind.

This is used for tools that are not yet officially released.

For example, the default release type for Kind::Fleet is Type::Preview.