Struct jb_lib::tool::Tool

pub struct Tool {
    pub kind: Kind,
    pub version: Option<Version>,
    pub directory: Option<PathBuf>,
}

A JetBrains tool

This struct represents a JetBrains tool, such as IntelliJ IDEA or PyCharm. It contains information about the tool, such as its name, version, and installation directory. However, it may not be installed yet.

Every field is optional, except for kind. However, they are all read-only and may only be set through the builder methods.

Examples

use jb_lib::tool::{Tool, Kind};

let tool = Tool::new(Kind::RustRover)
  .with_version("2021.2.1".parse().unwrap())
  .with_directory("/home/user/.local/share/JetBrains".into());

Fields
kind: Kind

The kind of tool

version: Option<Version>

The version of the tool

directory: Option<PathBuf>

The directory the tool is installed to or will be installed to

Implementations

impl Tool

pub fn new(kind: Kind) -> Self

pub fn list(directory: Option<PathBuf>) -> Result<Vec<Tool>>

List all installed JetBrains tools.

This function returns a list of all installed JetBrains tools.

Examples

use jb_lib::tool::Tool;
let tools = Tool::list(None).unwrap();

Errors

This function may return an error if it fails to read the installation directory, or any of the installed tools.

Panics

This function may panic if it fails to strip the tool name prefix from the tool directory or convert the path to a string.

pub fn with_version(self, version: Version) -> Self

Set the version of the tool.

Examples

use jb_lib::tool::{Tool, Kind};
let tool = Tool::new(Kind::RustRover)
 .with_version("2021.2.1".parse().unwrap());

pub fn with_directory(self, directory: PathBuf) -> Self

Set the installation directory of the tool.

Examples

use jb_lib::tool::{Tool, Kind};
let tool = Tool::new(Kind::RustRover)
.with_directory("/home/user/.local/share/JetBrains".into());

pub fn is_linked(&self) -> bool

Check if the tool is linked.

This function checks if the tool is the one linked to the icon directory.

Examples

use jb_lib::tool::{Tool, Kind};
let tool = Tool::new(Kind::RustRover)
  .with_version("2021.2.1".parse().unwrap());

if tool.is_linked() {
   println!("Rust Rover is linked");
}

pub fn link(&self) -> Result<()>

Link the tool.

This function links the tool to the icon directory and the binary directory. It will attempt to delete any existing symlinks, regardless of whether they exist or not (to avoid errors when linking).

Errors

This function may return an error if the tool is already linked. It may also return an error if it fails to create the icon or binary symlink.

Examples

use jb_lib::tool::{Tool, Kind};
let tool = Tool::new(Kind::RustRover)
 .with_version("2021.2.1".parse().unwrap());

match tool.link() {
   Ok(_) => println!("Linked Rust Rover"),
  Err(e) => println!("Failed to link Rust Rover: {}", e),
}

pub fn unlink(&self) -> Result<()>

Unlink the tool.

This function unlinks the tool from the icon directory and the binary directory. It will attempt to delete any existing symlinks, regardless of whether they exist or not (to avoid errors when unlinking).

However, it will try to find an alternative version to link to, if one exists.

Errors

This function may return an error if the tool is not linked. It may also return an error if it fails to remove the icon or binary symlink, as well as if it fails to create the link to the alternative version (if any).

Examples

use jb_lib::tool::{Tool, Kind};
let tool = Tool::new(Kind::RustRover)
  .with_version("2021.2.1".parse().unwrap());

match tool.unlink() {
  Ok(_) => println!("Unlinked Rust Rover"),
 Err(e) => println!("Failed to unlink Rust Rover: {}", e),
}

pub fn name(&self) -> String

Get the name of the tool.

This function returns the name of the tool, including its version. This is used for the name of the tool’s directory.

Examples

use jb_lib::tool::{Tool, Kind};
let tool = Tool::new(Kind::RustRover)
 .with_version("2021.2.1".parse().unwrap());

println!("Tool name: {}", tool.name());

pub fn as_path(&self) -> PathBuf

Get the path to the tool.

This function returns the path to the tool’s directory.

Examples

use jb_lib::tool::{Tool, Kind};
let tool = Tool::new(Kind::RustRover)
  .with_version("2021.2.1".parse().unwrap());

println!("Tool path: {}", tool.as_path().display());

pub fn download_link(&self) -> Result<Download>

Get the download link for the tool.

This function returns the download link for the tool, including the one for the checksum and the size of the download. If no version is specified, it will return the latest version.

Errors

This function may return an error if it fails to fetch the releases for the tool, or if it fails to parse the releases. It may also return an error if it fails to find a compatible download with the given version.

Examples

use jb_lib::tool::{Tool, Kind};
let tool = Tool::new(Kind::RustRover);

println!("Download link: {}", tool.download_link().unwrap().link);

pub fn install(&mut self) -> Result<()>

Install the tool.

This function installs the tool to the given directory. If no directory is specified, it will install to the default directory.

This function takes a mutable reference to the tool, as it will update the version of the tool to the one that was installed.

Errors

This function may return an error if it fails to create the tool directory, or if the tool is already installed. It may also return an error if it fails to download the tool, or if it fails to extract the archive, as well as if it fails to link the tool.

Panics

This function may panic if it fails to get the filename from the download link.

Examples

use jb_lib::tool::{Tool, Kind};
let mut tool = Tool::new(Kind::RustRover)
 .with_version("2021.2.1".parse().unwrap());

match tool.install() {
   Ok(_) => println!("Installed Rust Rover"),
  Err(e) => println!("Failed to install Rust Rover: {}", e),
}

pub fn uninstall(&self) -> Result<()>

Uninstall the tool.

This function uninstalls the tool from the given directory. If no directory is specified, it will uninstall from the default directory.

Errors

This function may return an error if it fails to remove the tool directory, or if the tool is not installed. It may also return an error if it fails to unlink the tool.

Examples

use jb_lib::tool::{Tool, Kind};

let tool = Tool::new(Kind::RustRover)
  .with_version("2021.2.1".parse().unwrap());

match tool.uninstall() {
   Ok(_) => println!("Uninstalled Rust Rover"),
  Err(e) => println!("Failed to uninstall Rust Rover: {}", e),
}